moved utils + help string + string_util
