+++ lib/simbricks/nicbm/nicbm.h
@@ -31,10 +31,10 @@
 #include <set>
 
 #include <simbricks/base/cxxatomicfix.h>
-#include <simbricks/trace/log.h>
 extern "C" {
 #include <simbricks/nicif/nicif.h>
 }
+#include "lib/utils/log.h"
 
 namespace nicbm {
